B-0: fix a float32 ordering regression, and stop overclaiming
Review of 337266f found one blocking defect and four overstatements. All
are accepted.

The blocking one. The extracted mixture wrapper built its tensors without
a dtype, so they defaulted to float32 while the legacy code used Python
doubles throughout. That is not a display-level difference: two
candidates whose true scores differ by less than a float32 step near 0.5
— about 6e-8 — collapse to an exact tie, which is then resolved by sort
stability, that is by input order, and can move a candidate across the
top-k boundary a clinician sees. The commit claimed behaviour
preservation and did not have it.

Fixed by specifying float64, which is verified to reproduce the legacy
Python arithmetic exactly rather than approximately: 2000 random cases
per eta, asserted equal with no tolerance. A tolerance there would permit
the rounding the test exists to forbid.

The regression tests use a near-tie pair constructed for the purpose
rather than the one the review supplied, because that one does not
actually collapse in float32 — a test built on it would pass with the bug
present. A separate test pins the constructed pair's collapsing property,
so if future rounding behaviour changes the suite says so instead of
quietly testing nothing. Ordering and top-k boundary are asserted
separately from arithmetic.

The overstatements. The module docstring said every surface that produces
a disease score composes these functions. It does not: the evaluator
still has its own implementation and is migrated later in B-0. Writing
the target state in the present tense is the documentation defect this
project has spent several rounds removing from other files, and it should
not have been introduced here. The claim is now scoped to what is true,
with the migration marked in progress.

The scope boundary is also recorded rather than left implicit: this
module cannot serve the training path, because .import-linter.ini places
src.inference above src.training and a training import would invert the
layering. Training keeps its own implementation and the correspondence is
a property for equivalence tests, not shared code.

Three narrower corrections. The availability mask currently separates
only "nothing to measure from" from "a distance was produced" — the other
unavailable states are still detected by the caller and collapsed to 0.0
— so the docstring no longer implies it carries the full B-1 contract.
sp_mean_distances is labelled batched in interface but not yet vectorised,
with the cost stated: about 4,000 slice scans at 200 candidates, over
550,000 at full-universe scale. And pool_patient_embeddings now builds its
index on the embedding tensor's device rather than relying on an implicit
CPU-index convenience that the full-universe evaluator would not enjoy.

Structural authority tests now cover every wrapper rather than only the
cosine path: both shortest-path primitives, the availability mask's
collapse to 0.0, the mixture, and the no-SP pure-GNN degradation.

Full suite 634 passed, 7 skipped. The eighteen integration tests that pin
the scoring path still pass unmodified.

Status: B-0 in progress, scoring-extraction slice only. Not measurement
ready. Vectorised lookup, evaluator migration, modes A-D, untruncated
metrics, configuration manifests and negative-set instrumentation all
remain.

+++ src/inference/pipeline.py
@@ -1322,10 +1322,17 @@ class DiagnosisPipeline:
 
         from src.inference.scoring import mix_embedding_and_sp_scores
 
+        # float64 is required, not incidental. The scores arriving here are Python
+        # doubles; constructing float32 tensors would round them to ~7 significant
+        # digits, and two candidates whose true scores differ below that threshold
+        # would collapse to an exact tie. A tie is then resolved by sort stability
+        # — that is, by input order — which can move a candidate across the top-k
+        # boundary a clinician sees. The extraction is only behaviour-preserving in
+        # double precision.
         combined = float(
             mix_embedding_and_sp_scores(
-                torch.tensor([emb_score]),
-                torch.tensor([sp_score]),
+                torch.tensor([emb_score], dtype=torch.float64),
+                torch.tensor([sp_score], dtype=torch.float64),
                 self.config.eta,
             )[0]
         )
